Extract JPS based IDE and tests into separate modules
