+++ intellij-plugin-structure/settings.gradle.kts
@@ -15,6 +15,7 @@ include("structure-ide-jps")
 include("structure-youtrack")
 include("structure-teamcity-recipes")
 include("tests")
+include("tests-jps")
 
 dependencyResolutionManagement {
   versionCatalogs {
