+++ intellij-plugin-structure/structure-ide-classes/src/main/java/com/jetbrains/plugin/structure/ide/classes/IdeResolverCreator.kt
@@ -4,25 +4,17 @@
 
 package com.jetbrains.plugin.structure.ide.classes
 
-import com.jetbrains.plugin.structure.base.utils.closeOnException
 import com.jetbrains.plugin.structure.base.utils.isDirectory
-import com.jetbrains.plugin.structure.base.utils.listFiles
 import com.jetbrains.plugin.structure.base.utils.listJars
-import com.jetbrains.plugin.structure.base.utils.simpleName
 import com.jetbrains.plugin.structure.classes.resolvers.CompositeResolver
-import com.jetbrains.plugin.structure.classes.resolvers.DirectoryResolver
 import com.jetbrains.plugin.structure.classes.resolvers.EMPTY_RESOLVER
 import com.jetbrains.plugin.structure.classes.resolvers.FileOrigin
 import com.jetbrains.plugin.structure.classes.resolvers.Resolver
 import com.jetbrains.plugin.structure.classes.resolvers.buildJarOrZipFileResolvers
 import com.jetbrains.plugin.structure.ide.Ide
-import com.jetbrains.plugin.structure.ide.IdeManagerImpl
-import com.jetbrains.plugin.structure.ide.IdeManagerImpl.Companion.isCompiledCommunity
-import com.jetbrains.plugin.structure.ide.IdeManagerImpl.Companion.isCompiledUltimate
 import com.jetbrains.plugin.structure.ide.IdeManagerImpl.Companion.isDistributionIde
 import com.jetbrains.plugin.structure.ide.InvalidIdeException
 import com.jetbrains.plugin.structure.ide.classes.resolver.ProductInfoClassResolver
-import com.jetbrains.plugin.structure.ide.getRepositoryLibrariesJars
 import com.jetbrains.plugin.structure.ide.layout.InvalidIdeLayoutException
 import java.nio.file.Path
 
@@ -34,13 +26,14 @@ object IdeResolverCreator {
   @JvmStatic
   fun createIdeResolver(readMode: Resolver.ReadMode, ide: Ide): Resolver = createIdeResolver(ide, IdeResolverConfiguration(readMode))
 
+  @Throws(InvalidIdeException::class)
   @JvmStatic
   fun createIdeResolver(ide: Ide, configuration: IdeResolverConfiguration): Resolver {
     val idePath = ide.idePath
-    return when {
-      isDistributionIde(idePath) -> getIdeResolverFromDistribution(ide, configuration)
-      isCompiledCommunity(idePath) || isCompiledUltimate(idePath) -> getIdeResolverFromCompiledSources(idePath, configuration.readMode, ide)
-      else -> throw InvalidIdeException(idePath, "Invalid IDE $ide at $idePath")
+    if (isDistributionIde(idePath)) {
+      return getIdeResolverFromDistribution(ide, configuration)
+    } else {
+      throw InvalidIdeException(idePath, "Invalid IDE $ide at $idePath")
     }
   }
 
@@ -74,36 +67,5 @@ object IdeResolverCreator {
       throw InvalidIdeException(ide.idePath, e)
     }
   }
-
-  //TODO: Resolver created this way contains all libraries declared in the project,
-  // including those that don't go to IDE distribution. So such a created resolver may
-  // resolve classes differently than they are resolved when running IDE.
-  // IDE sources can generate so-called "project-structure-mapping.json", which contains mapping
-  // between compiled modules and jar files to which these modules are packaged in the final distribution.
-  // We can use this mapping to construct a true resolver without irrelevant libraries.
-  private fun getIdeResolverFromCompiledSources(idePath: Path, readMode: Resolver.ReadMode, ide: Ide): Resolver {
-    val resolvers = arrayListOf<Resolver>()
-    resolvers.closeOnException {
-      resolvers += getJarsResolver(idePath.resolve("lib"), readMode, IdeFileOrigin.SourceLibDirectory(ide))
-      resolvers += getRepositoryLibrariesResolver(idePath, readMode, ide)
-
-      val compiledClassesRoot = IdeManagerImpl.getCompiledClassesRoot(idePath)!!
-      compiledClassesRoot.listFiles().forEach { moduleRoot ->
-        val fileOrigin = IdeFileOrigin.CompiledModule(ide, moduleRoot.simpleName)
-        resolvers += DirectoryResolver(moduleRoot, fileOrigin, readMode)
-      }
-
-      if (isCompiledUltimate(idePath)) {
-        resolvers += getJarsResolver(idePath.resolve("community").resolve("lib"), readMode, IdeFileOrigin.SourceLibDirectory(ide))
-      }
-      return CompositeResolver.create(resolvers)
-    }
-  }
-
-  private fun getRepositoryLibrariesResolver(idePath: Path, readMode: Resolver.ReadMode, ide: Ide): Resolver {
-    val jars = getRepositoryLibrariesJars(idePath)
-    return CompositeResolver.create(buildJarOrZipFileResolvers(jars, readMode, IdeFileOrigin.RepositoryLibrary(ide)))
-  }
-
 }
