+++ intellij-plugin-structure/structure-ide-jps/build.gradle.kts
@@ -1,5 +1,7 @@
 dependencies {
-  api(project(":structure-intellij"))
+  api(project(":structure-classes"))
+  api(project(":structure-ide"))
+  api(project(":structure-ide-classes"))
 
   implementation(libs.platform.util)
   implementation(libs.platform.jps.model.core)
