+++ intellij-plugin-structure/structure-ide-jps/src/main/resources/META-INF/services/com.jetbrains.plugin.structure.ide.IdeManager
@@ -1,0 +1,1 @@
+com.jetbrains.plugin.structure.ide.jps.CompiledIdeManager
