+++ intellij-plugin-structure/tests-jps/build.gradle.kts
@@ -1,0 +1,6 @@
+dependencies {
+  testImplementation(sharedLibs.junit)
+  testImplementation(project(":structure-ide"))
+  testImplementation(project(":structure-ide-jps"))
+  testImplementation(project(path = ":tests", configuration = "testOutput"))
+}
